feat(OpossumReporter): Always exclude "rootless" packages
Packages which are excluded by a scope exclude get excluded from the
report. Such packages are not contained in the project's release
artifact. As packages which are not referenced from any project also
are not contained in the release artifact, they also should be excluded
from the report for consistency.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>